refactor(gulp): Ordenando estructura de ejecución de tareas gulp.
