+++ frontend/miproyecto/CHANGELOG.md
@@ -1,3 +1,35 @@
+### 1.0.3 (2014-11-13)
+
+
+#### Bug Fixes
+
+* **jadeflux:** Se corrige error compilacion en versiones de jade > 1.3.1. ([94b0428e](https://github.com/jansanchez/flux.git/commit/94b0428e88fee58ae7b510a586517cdc2706d98f))
+
+
+#### Features
+
+* **flux:** Actualizando Flux a su version 1.0.2. ([7a9cc7ba](https://github.com/jansanchez/flux.git/commit/7a9cc7ba4ec8f92f905312446ce6744acd8744d8))
+* **gulp:**
+  * Agregando tareas para generar fuente de letras automatico. ([fc1946d4](https://github.com/jansanchez/flux.git/commit/fc1946d4510b6c407c1ebd83397eecb65d74ed11))
+  * Agregando tareas para generar fuente de iconos automatico. ([f8a61257](https://github.com/jansanchez/flux.git/commit/f8a612579963fe8b679c241cc5bbc8b00026458a))
+
+
+### 1.0.2 (2014-11-13)
+
+
+#### Bug Fixes
+
+* **jadeflux:** Se corrige error compilacion en versiones de jade > 1.3.1. ([94b0428e](https://github.com/jansanchez/flux.git/commit/94b0428e88fee58ae7b510a586517cdc2706d98f))
+
+
+#### Features
+
+* **flux:** Actualizando Flux a su version 1.0.2. ([7a9cc7ba](https://github.com/jansanchez/flux.git/commit/7a9cc7ba4ec8f92f905312446ce6744acd8744d8))
+* **gulp:**
+  * Agregando tareas para generar fuente de letras automatico. ([fc1946d4](https://github.com/jansanchez/flux.git/commit/fc1946d4510b6c407c1ebd83397eecb65d74ed11))
+  * Agregando tareas para generar fuente de iconos automatico. ([f8a61257](https://github.com/jansanchez/flux.git/commit/f8a612579963fe8b679c241cc5bbc8b00026458a))
+
+
 ### 1.0.1 (2014-11-06)
